+++ tests/aws_jobs_json/prj.conf
@@ -1,0 +1,4 @@
+CONFIG_ZTEST=y
+CONFIG_JSON_LIBRARY=y
+CONFIG_ZTEST_STACKSIZE=2048
+CONFIG_HEAP_MEM_POOL_SIZE=4096
